Rearranging of files and updated view.
Package structure changed slightly for the App, Game, View (renamed to ChessView). Currently src is listed twice in the directory, this will need to be fixed but Eclipse was complaining about the way it was imported from IntelliJ. 

I cleaned up the view a bit, added an HBox for buttons (with only the concede button there at the moment, more to be added later). I ended up commenting out a lot of the CSS styling - I think we should work primarily with Javafx until we have a working product, then we can clean it up and add in the small things like that. Right now it's just visual clutter.